requirements.txt
torch>=2.5.0
mamba-ssm>=2.2.0
causal-conv1d>=1.4.0
datasets>=4.0.0
transformers>=5.0.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
numpy>=1.24.0
pyyaml>=6.0
